move parameter validation logic into BlockConstraints
