+++ apps/omg_watcher_rpc/test/omg_watcher_rpc/web/validators/block_constraints_test.exs
@@ -17,6 +17,12 @@ defmodule OMG.WatcherRPC.Web.Validator.BlockConstraintsTest do
 
   alias OMG.WatcherRPC.Web.Validator.BlockConstraints
 
+  @valid_block %{
+    "hash" => "0x" <> String.duplicate("00", 32),
+    "number" => 1000,
+    "transactions" => ["0x00"]
+  }
+
   describe "parse/1" do
     test "returns page and limit constraints when given page and limit params" do
       request_data = %{"page" => 1, "limit" => 100}
@@ -56,6 +62,40 @@ defmodule OMG.WatcherRPC.Web.Validator.BlockConstraintsTest do
     test "returns a :validation_error when the given limit is not an integer" do
       assert BlockConstraints.parse(%{"page" => 3.14}) == {:error, {:validation_error, "page", :integer}}
       assert BlockConstraints.parse(%{"page" => "abcd"}) == {:error, {:validation_error, "page", :integer}}
+    end
+  end
+
+  describe "parse_to_validate/1" do
+    test "rejects invalid Merkle root hash" do
+      invalid_hash = "0x1234"
+      invalid_block = Map.replace!(@valid_block, "hash", invalid_hash)
+
+      assert {:error, {:validation_error, "hash", {:length, 32}}} ==
+               BlockConstraints.parse_to_validate(invalid_block)
+    end
+
+    test "rejects non-list transactions parameter" do
+      invalid_transactions_param = "0x1234"
+      invalid_block = Map.replace!(@valid_block, "transactions", invalid_transactions_param)
+
+      assert {:error, {:validation_error, "transactions", :list}} ==
+               BlockConstraints.parse_to_validate(invalid_block)
+    end
+
+    test "rejects non-hex elements in transactions list" do
+      invalid_tx_rlp = "0xZ"
+      invalid_block = Map.replace!(@valid_block, "transactions", [invalid_tx_rlp])
+
+      assert {:error, {:validation_error, "transactions.hash", :hex}} ==
+               BlockConstraints.parse_to_validate(invalid_block)
+    end
+
+    test "rejects invalid block number parameter" do
+      invalid_blknum = "ONE THOUSAND"
+      invalid_block = Map.replace!(@valid_block, "number", invalid_blknum)
+
+      assert {:error, {:validation_error, "number", :integer}} ==
+               BlockConstraints.parse_to_validate(invalid_block)
     end
   end
 end
